10.1.5 structs

+++ Source/Forged.MapServer/DataStorage/Structs/A/AreaTableRecord.cs
@@ -22,7 +22,7 @@ public sealed class AreaTableRecord
     public uint UwIntroSound;
     public byte FactionGroupMask;
     public float AmbientMultiplier;
-    public byte MountFlags;
+    public uint MountFlags;
     public short PvpCombatWorldStateID;
     public byte WildBattlePetLevelMin;
     public byte WildBattlePetLevelMax;
